changes
final changes from my part, both in the code and in the requirements.
