+++ Requirements.txt
@@ -1,27 +1,19 @@
 alpha-vantage==2.3.1
-beautifulsoup4==4.11.1
+beautifulsoup4==4.12.2
 bs4==0.0.1
+DateTime==5.1
 finnhub-python==2.4.18
-numpy==1.21.5
-pandas==1.4.4
-pandas-datareader==0.10.0
-plotly==5.9.0
-plotly.express==0.4.1
-requests==2.28.1
-requests-file==1.5.1
+numpy==1.24.3
+pandas==2.0.1
+plotly==5.14.1
+plotly-express==0.4.1
+requests==2.31.0
 requests-html==0.10.0
-requests-oauthlib==1.3.1
-scikit-image==0.19.2
-scikit-learn==1.0.2
-scikit-learn-intelex==2021.20221004.171935
-streamlit==1.20.0
-streamlit-option-menu==0.3.2
+scikit-learn==1.2.2
+scipy==1.10.1
+sklearn==0.0.post5
+streamlit==1.22.0
+streamlit-option-menu==0.3.4
 xgboost==1.7.5
 yahoo-fin==0.8.9.1
-yfinance==0.2.12
-datetime
-plotly.graph_objs
-sklearn
-yahoo_fin
-yahoo_fin.stock_info 
-alpha_vantage.fundamentaldata
+yfinance==0.2.18
